Move identity tables to their own package
