+++ src/main/java/uk/ac/ucl/rits/inform/informdb/identity/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package stores Entity classes to map the core Emap-Star tables the describe patient identity.
+ */
+
+package uk.ac.ucl.rits.inform.informdb.identity;
