+++ src/main/java/uk/ac/ucl/rits/inform/informdb/package-info.java
@@ -1,5 +1,5 @@
 /**
-* This package stores Entity classes to map Inform-DB tables.
+* This package stores classes that are common to all parts of the EMAP-Star schema
 */
 
 package uk.ac.ucl.rits.inform.informdb;
